Addition of telethon package
The default python-telegram-bot api does not allow the bot to retrieve historic
messsages in a group. The telethon package does have a way to do this but it can
only do so using an actual user account as certain api functionality is blocked
for bot accounts.

Two possible fixes:

1. Create a separate user account for the KevinMaloneBot. This would circumvent
the telethon restrictions.

2. Make the KevinMaloneBot a group admin, allowing it to receive ALL messages
sent in the group. Will need to store them in a separeate db so they can be
retrieved when it is time to summarize. Will require the bot to be up at all
times otherwise it won't be able too receive messages.