Trialing of chat message retrieval
Completed the testing of the chat message retrieval from a group chat based on
an incoming messsage prompt.

Next step will be to build the actual components required in a proper script(as
opposed to a test file).

Also added an initial csv file to whitelist chats that are permitted to submit
requests for summarization. Will need to implement that checking layer.